team_service/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
python-decouple>=3.8
djangorestframework-simplejwt>=5.3.0 
psycopg2-binary
dotenv
